Properly allow the x-amz-date header to override Date
The S3 developer guide shows that the Date header should be omitted when
the x-amz-date header is provided:

https://s3.amazonaws.com/doc/s3-developer-guide/RESTAuthentication.html

This patchset removes x-amz-date from the list of date-like things to
check (e.g. Expires param, Date header) and overrides these when
present. The result for us is that we can use multipart signed documents
properly with EvaporateJS.